bytes: make iobuf hash operations opt-in
Consumer of the iobuf headers were forced to include numerous headers
that most iobuf users did not need. One of these was iobuf hashing
operations. This commit splits out these into an opt-in header and
updates users.
