+++ src/v/bytes/tests/iobuf_utils_tests.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "bytes/hash.h"
 #include "bytes/iostream.h"
 #include "bytes/tests/utils.h"
